Rename file to meet the format specifications
