+++ plugins/main/public/components/overview/vulnerabilities/common/components/document-view-table-and-json.tsx
